Add CI for plone.base.
We should catch it when we miss a dependency, or notice when a dependency pulls in too many others.
We currently miss the `plone.schema` dependency, needed for the `Email` field.
If we add it, it pulls in `plone.app.z3cform` for `IPloneFormLayer`, and this package pulls in lots of `plone.app` packages.