Fix traversal during get_top_site_from_url.
My previous fix broke several tests.  This appears to be the result of trying to
traverse iteratively through the structure of a site (which my code does and the
previous code does not do in the same way), and encountering `Unauthorized`
exceptions during traversal (which is a legitimate thing to get, as /a/b/c may
be accessible to user XXX, but /a/b may not necessarily be).

The fix is to use unrestricted traversal (bypassing security), and then to return
the selected object by using *restricted* traversal in that case.

I am confident this should not result in a security issue — however I ask the
true expert Plonistas to verify this confidence is warranted.  Thanks.

And sorry for the breakage.  We need CI tests in `plone.base`.